+++ reference/sources/beancount/code/core/amount.py
@@ -123,9 +123,9 @@ class Amount(NamedTuple("Amount", [("number", Optional[Decimal]), ("currency", s
         Returns:
           A new instance of Amount, with the negative number of units.
         """
-        assert isinstance(
-            self.number, Decimal
-        ), "Amount's number is not a Decimal instance: {}".format(self.number)
+        assert isinstance(self.number, Decimal), (
+            "Amount's number is not a Decimal instance: {}".format(self.number)
+        )
         return Amount(-self.number, self.currency)
 
     @staticmethod
@@ -139,9 +139,7 @@ class Amount(NamedTuple("Amount", [("number", Optional[Decimal]), ("currency", s
         Returns:
           A new instance of Amount.
         """
-        match = re.match(
-            r"\s*([-+]?[0-9.]+)\s+({currency})".format(currency=CURRENCY_RE), string
-        )
+        match = re.match(r"\s*([-+]?[0-9.]+)\s+({currency})".format(currency=CURRENCY_RE), string)
         if not match:
             raise ValueError("Invalid string for amount: '{}'".format(string))
         number, currency = match.group(1, 2)
@@ -174,12 +172,10 @@ def mul(amount: Amount, number: Decimal) -> Amount:
     Returns:
       An Amount, with the same currency, but with 'number' times units.
     """
-    assert isinstance(
-        amount.number, Decimal
-    ), "Amount's number is not a Decimal instance: {}".format(amount.number)
-    assert isinstance(number, Decimal), "Number is not a Decimal instance: {}".format(
-        number
+    assert isinstance(amount.number, Decimal), (
+        "Amount's number is not a Decimal instance: {}".format(amount.number)
     )
+    assert isinstance(number, Decimal), "Number is not a Decimal instance: {}".format(number)
     return Amount(amount.number * number, amount.currency)
 
 
@@ -192,12 +188,10 @@ def div(amount: Amount, number: Decimal) -> Amount:
     Returns:
       An Amount, with the same currency, but with amount units divided by 'number'.
     """
-    assert isinstance(
-        amount.number, Decimal
-    ), "Amount's number is not a Decimal instance: {}".format(amount.number)
-    assert isinstance(number, Decimal), "Number is not a Decimal instance: {}".format(
-        number
+    assert isinstance(amount.number, Decimal), (
+        "Amount's number is not a Decimal instance: {}".format(amount.number)
     )
+    assert isinstance(number, Decimal), "Number is not a Decimal instance: {}".format(number)
     return Amount(amount.number / number, amount.currency)
 
 
@@ -211,12 +205,12 @@ def add(amount1: Amount, amount2: Amount) -> Amount:
       An instance of Amount, with the sum the two amount's numbers, in the same
       currency.
     """
-    assert isinstance(
-        amount1.number, Decimal
-    ), "Amount1's number is not a Decimal instance: {}".format(amount1.number)
-    assert isinstance(
-        amount2.number, Decimal
-    ), "Amount2's number is not a Decimal instance: {}".format(amount2.number)
+    assert isinstance(amount1.number, Decimal), (
+        "Amount1's number is not a Decimal instance: {}".format(amount1.number)
+    )
+    assert isinstance(amount2.number, Decimal), (
+        "Amount2's number is not a Decimal instance: {}".format(amount2.number)
+    )
     if amount1.currency != amount2.currency:
         raise ValueError(
             "Unmatching currencies for operation on {} and {}".format(amount1, amount2)
@@ -234,12 +228,12 @@ def sub(amount1: Amount, amount2: Amount) -> Amount:
       An instance of Amount, with the difference between the two amount's
       numbers, in the same currency.
     """
-    assert isinstance(
-        amount1.number, Decimal
-    ), "Amount1's number is not a Decimal instance: {}".format(amount1.number)
-    assert isinstance(
-        amount2.number, Decimal
-    ), "Amount2's number is not a Decimal instance: {}".format(amount2.number)
+    assert isinstance(amount1.number, Decimal), (
+        "Amount1's number is not a Decimal instance: {}".format(amount1.number)
+    )
+    assert isinstance(amount2.number, Decimal), (
+        "Amount2's number is not a Decimal instance: {}".format(amount2.number)
+    )
     if amount1.currency != amount2.currency:
         raise ValueError(
             "Unmatching currencies for operation on {} and {}".format(amount1, amount2)
@@ -255,9 +249,9 @@ def abs(amount: Amount) -> Amount:
     Returns:
       An instance of Amount.
     """
-    assert isinstance(
-        amount.number, Decimal
-    ), "Amount's number is not a Decimal instance: {}".format(amount.number)
+    assert isinstance(amount.number, Decimal), (
+        "Amount's number is not a Decimal instance: {}".format(amount.number)
+    )
     return amount if amount.number >= ZERO else Amount(-amount.number, amount.currency)